Feat(Mosaic): Calculate each cell's entropy
Each cell in the mosaic starts with the entire tileset as it's pool of
options of which tile to collapse to.

For each cell, we limit it's neighbours options pool by checking if any
of it's possible tiles are compatible with the original cell. If not,
discard it.

To verify if two tiles are compatible, we compare if the edge of one
cell agains't the reversed edge of the particular neighbor is the same.

At the end of the process each cell's option pool was limited by each of
it's 4 neighbours, thus calculating each cell's entropy.