Refactor(Mosaic): merge grid and gridOptions
At the start of the project, my state variable grid would only store
which cells were collapesed with which tiles, and null in cases where the
cell was not collapsed at all. gridOptions on the other side would store
all the possible tiles for each cell and be kept in sync with grid to do
so.

By eliminating gridOptions and storing all the possible tiles for each
cell in a state variable I'm am able to see if a variable is collapsed
(grid[cell].length === 1), not collapsed (grid[cell].length > 1), or if
a paradox was reached and there's no compatible tile to put in the cell
in order to solve the Mosaic (grid[cell].length === 0).

A fun side-effect of eliminating this data duplication is that at each
step the mosaic will render all the new collapsed cells at once, after
all the component now assumes that all cells with only one possible tile
are cells collapsed to that one tile. (Accidental feature, not a bug)