Feat(Mosaic): Propagate entropy on cell collapse
Previously, when a cell was collapsed it would limit only it's adjacent
cells possibilities for tiles. This commit takes a step further and
recalculates the entropy of the neighbors' neighbors (and so on) taking
in consideration the new value for the neighbors entropy. In effect, it
propagates these effects from the collapsed cell outwards.

Since 954526e (Refactor(Mosaic): merge grid and gridOptions,
2023-05-22), cells are considered collapsed when they have a entropy of
one, instead of the previous solution of flagging said cell as
collapsed. This introduced a bug where cells collapsed through direct
propagation wouldn't in turn propagate this change to their neighbors,
leading to cases where cells would collapse to wrong tiles (they
wouldn't match their neighbors). This commit fixes that.

It's worth noticing that limitCellsOptions() was modified into
getCongruentTiles(). In this new version instead of finding the
compatible tiles of an array against a match, it returns any tile that
matches any tile in a pattern. This change was added because now
uncollapsed cells also propagate entropy.