Refactor: Use positive values in CSS rotations
In CSS, positive <angle> values correspond to clockwise motion as
opposed to my initial assumption in which they would correspond to
anticlockwise motions as in math in general.

Because it is more intuitive to work with positive values I've changed
how tilesets are defined. Formerly, each tile's edges (and each edge's
sockets) were defined first from the right and then in an anticlockwise
motion. Now each tile starts defining it's edges (and sockets) from the
top and then go around in a clockwise motion.

In turn, this made me change the order of indexes returned in
getNeighborsIndexes() accordingly.