Feat(Mosaic): Backtrack to previous grid snapshots
For each time a cell is collapsed, the snapshot of the updated grid is
added to history.

Since each collapse is at best a guess, there was also added the
capability to return to a previous snapshot in cases where entropy
propagations from a cell collapse would result in other cells having
zero entropy (i.e. no possible compatible tiles to create the resulting
mosaic)