Feat: Add controls to columns and rows ammount
Instead of setting the ammount of columns and rows through the Mosaic
props, now it is possible to control those values directly through
buttons in the UI that responds to clicks and click & holds (touchscreen
included).

The mosaic component in turn now has an effect that resets it's state
completely when the columns and rows change. TileSize is not floored
anymore in order to make the mosaic aligh corretly with the buttons on
screens with small width.