Feat: Add grid dimensions as params to /generate
The /generate route loader will always receive the grid's amount of
columns and rows as URL parameters or redirect to itself with default
values for both when none is provided. Both input fields and the grid
are kept in sync with those values.

Every time the input fields change the react-router Form in sent,
thusly updating the grid's dimensions.