Feat: Add WrappingImage component
This component generates two canvases, one off-screen which holds the
final state of the Mosaic grid, and another that uses the former as a
pattern.

When the mosaic is finally finished, the second canvas is rendered on top
of it, and then off-setted and repainted every 15ms on both axis by one
pixel. Because the edges of the circuit mosaic are always the same, the
scrolling effect keeps on indefinitely.

The mosaicIsGenerating boolean state was then converted to mosaicStatus
with 3 possible values to ensure the WrappingImage component is only
rendered when the mosaic is complete