Refactor(WrappingImage): Update implementation
Instead of creating a canvas pattern and offsetting and repainting it
every so often, now a image is created based on that pattern, repeating
it in a 3x3 grid. This 3x3 grid contains all the possible offsets of the
pattern.

At first only it's center is shown to the user covering the Mosaic,
using absolute positioning, and then it's top and left CSS properties
are updated at 60fps.

Just testing the possibility of this being more efficient than the
canvas pattern solution.