Fix: Remove cells flicker when updating columns
Bug description: before showing the correct amount of cells when
updating the columns, the grid would flicker with -rows cells when
decreasing the numbers of rows, and extra rows when increasing. This
would happen because history is a state dependent on the state of
columns and rows to dictate its grid's size, and would rely on a
useEffect to reset itself with the proper number of cells on its grid.

Because effects run after renders the grid would flicker with the wrong
number of cells during the first render and then the effect would
trigger the corrected render.

Bug fix: Changing useEffect to useLayoutEffect triggers the history
state change before the first render, removing the flicker.

useLayoutEffect should be used sparingly because it hurts performance.

Also, state variables shouldn't depend on each other but I still haven't
figured it out yet ok? ok