Fix: Allow grids with 1 col and/or 1 row
Previously it wasn't possible to set a Mosaic with 1 row and/or 1
column. This limitation was set because when the initialHistory was
generated, it was always assumed that all cells could have all tiles as
possible options to collapse to.

Mosaics with one column and/or 1 row are special in this tile-set with
wrapping border, because a cell's neighbor could be itself. Calculating
the initial entropy of the grid then needs a extra step: To remove all
tiles which are not compatible with themselves in cells whose neighbors
are themselves (quite a mouthful huh).

This fix shouldn't work with more complex tile-sets, where the rules are
not calculated based on adjacency like sudoku. This project just support
one tile-set anyway. And it's not like a sudoku tile-set would make any
sense with wrapping borders. :p