Fix(generateButton): Not lose focus when disabled
Previous to this commit, navigating to the generateButton and triggering
it through the keyboard would result in the button losing focus after
being disabled. This behavior is undesirable because it forces the user
to have the button refocused in order to generate a new Mosaic.

In order to fix this I moved the responsability of disabling the button
from the disabled prop to the event handler itself.