Refactor: Change cols/rows through useSearchParams
The previous method consisting in using Form, inputs and useSubmit was
hacky and confusing. For once, Form had to be outside the definition of
the StatefulInput component. Both inputs had to be restyled and readonly
and even so they were focusable when navigating through the keyboard.

Using useSearchParams eliminates all of these inconveniences. There's no
need to have a Form nor inputs. The cols and rows values are received
and shown in the DOM and pressing the + and - buttons result in the
corresponding searchParam being updated. Concise and clear.