Refactor(WrappingImage): Infer tile src from mods
Previously it was necessary to describe the path to the tiles both in
the glob import and the img elements. Now, when loading the tiles, the
img elements src and their place in the tiles array are derived from the
modules returned from the glob import.

In the future there will be support for multiple tile-sets and this will
reduce the surface area for bugs related to changing a path and
forgetting to update the other.