Fix(About): Scroll screen to top after link click
On mobile when reading an About section and then switching to
another one with a smaller height the viewport position won't scroll
back to the top. So if you're deep on Acknowledgement and then switch to
any other About sub-route you can't read the subsection from
the beginning without manually scrolling back to the top.

This commit fixes this issue by scrolling the window to the top on every
link click on the About list.