Feat(WrappingImage): Turn scrolling manual
Some people have complained that having the final image to infinite
scroll just as it is generated made them dizzy. In the effort of
accessibility, I've deleted this feature and instead implemented one
in which the user can drag the final image with his pointer, which also
can be done indefinitely similar to the infinite scrolling.