Refactor: Enhance sync of derived state
The history state variable is an object with a field "grid" which length
is rows * columns, also state variables. It's not advisable in the React
docs to have state dependent on other states, alternatively it's
recommended to derive those states at runtime. This is not an option
here as we can't derive exactly which multiplication of rows and columns
would result in the lenght of the grid. It's a 1D data structure
representing a 2D layout after all.

The best we can do is to change those states together as closely as
possible, to avoid any de-sync. Because rows and columns are states
managed by React Router, the best place to sync the history state
variable with them is in the Generator's loader function, as any update
to rows and columns would re-render the Generator page, but not before
running its loader function.

The loader then returns an initalHistory object which is used as the
initial state of history (duh). Mosaic's useLayoutEffect was moved up to
Generator to reflect those changes and resetHistory() was discarded
completely because of that.

GenerateButton, the last component that called on resetHistory()
replaced its functionality with a simple setHistory call.