Feat(core): Add weight support for tilesets
Now when defining tilesets, one can specify different weights for each
tile in order to make one more or less probable of a cell being
collapsed to. The function collapseCell was added to consider the tiles
weights when collapsing a cell.

Now that tiles don't necessarily have the same weights, a cell's entropy
is not the amount of tiles in its domain anymore. In order to keep using
the "least entropy" heuristic the function getCellEntropy was added. It
calculates any cell's Shannon Entropy and maps it to values between 0
and 1. The lesser the value, the lesser the entropy/uncertainty of the
cell.

Now generated images look a lot more like actual circuit boards c: