Feat: Migrate from edge compatibility to whitelist
Previously this implementation of WFC used to read edge labels from all
tiles in a tileset and would then compare between them in order to find
which tiles were compatible between themselves. I've used this method
because that's what Dan Shiffman has used in his implementation. It
made sense back then because he would only support the Circuit tileset,
and for that one edge labels with a simple blacklist would suffice.

But there are many other tilesets and not all of them have compatibility
rules that emerge based on their tile edges. The current method is
flawed because it can't express them. A whitelist is a more verbose
solution but it doesn't have this limitation. So I've adapted my
blacklist implementation into a whitelist. This is good because I just
need to describe left and right tile pair rules instead of all 4
directions.

Maxin Gumin was able to shorten his whitelists even more by adding a
symmetry label to each tile corespondent to actions of the dihedral
group D4. With this, many whitelist rules can be deduced by checking the
symmetry of other permitted tile pairs. At least that's all I've could
glimpse from reading his code, but it's still very difficult for me to
understand, so I don't know if I'll ever get to implement this. Who
knows? The future is a mystery...