Feat: Add credits sections
All people who granted me permission to use their tilesets should be
credited. This commit adds a README section exclusive for them as well
as a section in the AboutAcknowledgments page. A lot of stuff that was
in this last component was moved to a new AboutAboutInspirations page,
it seems more fitting this way.