Fix: Correct the getCellEntropy function
Previously the function would use the cell variable instead of
validTiles when calculating the sumOfWeights and when scaling entropy
between 0 and 1. That implementation was flawed and it was easy to
perceive it on tilesets with no tiles with different weights, as cells
in those generations wouldn't never have their entropy reduced.

Now, the only cell with a entropy of zero is an empty cell, and a cell
with only one tile, although its collapsed automatically, can have an
entropy only as low as the probability of randomly picking the tile with
lowest weight in the tileset.

This fix brings constraint propagation back since commit bd2bb41
(Feat(core): Add weight support for tilesets, 2023-12-13), and makes it
so much easier to generate images with the Castle tileset.